fix(scanner): map stdint enum underlying types (uint32_t → uint32, not int64)
Clang reports an enum's fixedUnderlyingType by its written spelling, so
CF_OPTIONS(uint32_t, …) / CF_ENUM(int32_t, …) yield qualType "uint32_t" /
"int32_t". objcIntTypeToGo only recognised the keyword spellings ("unsigned
int", "int", …), so every stdint-typed enum fell through to the int64
default and was emitted 8 bytes wide.

That mislaid every struct embedding such an enum: e.g. CMTimeFlags is
CF_OPTIONS(uint32_t) but was int64, so CMTime came out 32 bytes instead of
C's 24 (flags at offset 16 instead of 12) — a latent ABI bug for any
by-value use, and it forced CMTime-containing structs off the typed path.

Add the stdint spellings to objcIntTypeToGo (mirroring the emitter's
mapEnumGoType). Re-scan corrects 2813 enums: ~2131 width fixes
(→uint32/uint16/uint8/int32/int16/int8) and 682 sign fixes (→uint64).
CMTimeFlags is now uint32 and CMTime is 24 bytes.

No overflow regressions (high-bit values fit the now-unsigned types;
UpgradeEnumTypeIfOverflow still applies). Parity 0 both pipelines;
deterministic; libraries build; tests pass.

+++ bindings/internal/raw/libraries/oslog/bridge/oslog_bridge.m
@@ -16,7 +16,7 @@ void * oslog_fn_os_log_create(const char * subsystem, const char * category, voi
 }
 
 // ID: objc-sym oslog.os_log_type_enabled
-bool oslog_fn_os_log_type_enabled(void * oslog, int64_t type_, void **outException) {
+bool oslog_fn_os_log_type_enabled(void * oslog, uint8_t type_, void **outException) {
     @autoreleasepool {
         @try {
             bool _result = (bool)os_log_type_enabled(oslog, type_);
@@ -29,7 +29,7 @@ bool oslog_fn_os_log_type_enabled(void * oslog, int64_t type_, void **outExcepti
 }
 
 // ID: objc-sym oslog._os_log_impl
-void oslog_fn__os_log_impl(void * dso, void * log, int64_t type_, const char * format, void * buf, uint32_t size, void **outException) {
+void oslog_fn__os_log_impl(void * dso, void * log, uint8_t type_, const char * format, void * buf, uint32_t size, void **outException) {
     @autoreleasepool {
         @try {
             _os_log_impl(dso, log, type_, format, buf, size);
@@ -40,7 +40,7 @@ void oslog_fn__os_log_impl(void * dso, void * log, int64_t type_, const char * f
 }
 
 // ID: objc-sym oslog._os_log_debug_impl
-void oslog_fn__os_log_debug_impl(void * dso, void * log, int64_t type_, const char * format, void * buf, uint32_t size, void **outException) {
+void oslog_fn__os_log_debug_impl(void * dso, void * log, uint8_t type_, const char * format, void * buf, uint32_t size, void **outException) {
     @autoreleasepool {
         @try {
             _os_log_debug_impl(dso, log, type_, format, buf, size);
@@ -51,7 +51,7 @@ void oslog_fn__os_log_debug_impl(void * dso, void * log, int64_t type_, const ch
 }
 
 // ID: objc-sym oslog._os_log_error_impl
-void oslog_fn__os_log_error_impl(void * dso, void * log, int64_t type_, const char * format, void * buf, uint32_t size, void **outException) {
+void oslog_fn__os_log_error_impl(void * dso, void * log, uint8_t type_, const char * format, void * buf, uint32_t size, void **outException) {
     @autoreleasepool {
         @try {
             _os_log_error_impl(dso, log, type_, format, buf, size);
@@ -62,7 +62,7 @@ void oslog_fn__os_log_error_impl(void * dso, void * log, int64_t type_, const ch
 }
 
 // ID: objc-sym oslog._os_log_fault_impl
-void oslog_fn__os_log_fault_impl(void * dso, void * log, int64_t type_, const char * format, void * buf, uint32_t size, void **outException) {
+void oslog_fn__os_log_fault_impl(void * dso, void * log, uint8_t type_, const char * format, void * buf, uint32_t size, void **outException) {
     @autoreleasepool {
         @try {
             _os_log_fault_impl(dso, log, type_, format, buf, size);
